arm7: merge SND_seq SND_unk
